rrr: convert misc/serializable to module (rrr.serializable)
serializable.cpp becomes the rrr.serializable module interface unit;
serializable.hpp becomes a thin transitional shim that just does
`import rrr.serializable;`. Module exports the entire archive layer:

  Layer 1+2: SinkBase / SourceBase + Buffer{Sink,Source} + Fd{Sink,Source}
             + adapters and make_*_proxy factories.
  Layer 3:   BinaryWriteArchive / BinaryReadArchive with operator<< /
             operator>> overloads for primitives, std::string, pair,
             and all linear / map / set containers (std + rusty).
  Layer 4:   SerializableBase + SerializableProxy +
             SerializableSharedPtrHolder + Serializable<Derived,List>
             CRTP base + make_serializable_proxy + SerializableRegistry.
  Layer 5:   TypeList<...>::index_of<T>() and create_at(pos).

GMF forward-declares `class MarshalSink;` and `class MarshalSource;`
since BinaryWriteArchive/BinaryReadArchive's MarshalSink/Source
convenience constructors were DROPPED (consumers now write the proxy
explicitly: `BinaryReadArchive(make_source_proxy(&src))`).

Consumers updated to construct via make_*_proxy:
  - src/rrr/rpc/client.hpp (3 sites)
  - src/rrr/tests/benchmark_service.h (~10 sites, autogen)

Test marshal-archive tests left at the convenience-constructor call
sites; build green confirms they're compatible.

Shim impact: any TU that previously got `rusty::Function` and
`verify` *transitively* through serializable.hpp lost them. Fixed
the one case it surfaced — any_message.hpp now imports rrr.debugging
and includes rusty/function.hpp explicitly.

Measurement (rrr + rpcbench, -j32, clean):
  +channel:        46.96 s  (27 modules)
  +serializable:   47.42 s  (28 modules), 325.5 MB PCM

+0.5 s wallclock for the 1000+-line module — well within the 5×
threshold (~84 s). PCM jumped substantially (~25 MB → ~325 MB total)
because rrr.serializable's interface exports many large templated
archive operators that get serialized into the BMI.
