+++ src/rrr/rpc/client.hpp
@@ -1730,7 +1730,7 @@ public:
         static_assert(std::is_invocable_v<F&, BinaryWriteArchive&>,
                       "request_with_options write_fn must accept BinaryWriteArchive&");
         MarshalSink sink(&serialized_args);
-        BinaryWriteArchive ar(&sink);
+        BinaryWriteArchive ar(make_sink_proxy(&sink));
         write_fn(ar);
         std::string args_bytes;
         size_t args_size = serialized_args.content_size();
