+++ src/rrr/tests/benchmark_service.h
@@ -447,7 +447,7 @@ private:
         {
             RpcFastPrimeRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.n;
             auto __typed_result__ = this->fast_prime(__typed_req__);
             auto sconn_opt = weak_sconn.upgrade();
@@ -471,7 +471,7 @@ private:
         {
             RpcFastDotProdRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.p1;
             __req_ar__ >> __typed_req__.p2;
             auto __typed_result__ = this->fast_dot_prod(__typed_req__);
@@ -496,7 +496,7 @@ private:
         {
             RpcFastAddRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.a;
             __req_ar__ >> __typed_req__.b;
             auto __typed_result__ = this->fast_add(__typed_req__);
@@ -521,7 +521,7 @@ private:
         {
             RpcFastNopRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.in_0;
             auto __typed_result__ = this->fast_nop(__typed_req__);
             auto sconn_opt = weak_sconn.upgrade();
@@ -544,7 +544,7 @@ private:
         {
             RpcFastVecRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.n;
             auto __typed_result__ = this->fast_vec(__typed_req__);
             auto sconn_opt = weak_sconn.upgrade();
@@ -568,7 +568,7 @@ private:
         {
             RpcPrimeRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.n;
             auto __typed_result__ = this->prime(__typed_req__);
             auto sconn_opt = weak_sconn.upgrade();
@@ -592,7 +592,7 @@ private:
         {
             RpcDotProdRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.p1;
             __req_ar__ >> __typed_req__.p2;
             auto __typed_result__ = this->dot_prod(__typed_req__);
@@ -617,7 +617,7 @@ private:
         {
             RpcAddRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.a;
             __req_ar__ >> __typed_req__.b;
             auto __typed_result__ = this->add(__typed_req__);
@@ -642,7 +642,7 @@ private:
         {
             RpcNopRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.in_0;
             auto __typed_result__ = this->nop(__typed_req__);
             auto sconn_opt = weak_sconn.upgrade();
@@ -665,7 +665,7 @@ private:
         {
             RpcAsyncNopRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.in_0;
             auto __async_req__ = std::move(req);
             auto __async_weak_sconn__ = weak_sconn;
@@ -691,7 +691,7 @@ private:
         {
             RpcSleepRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.sec;
             auto __typed_result__ = this->sleep(__typed_req__);
             auto sconn_opt = weak_sconn.upgrade();
@@ -714,7 +714,7 @@ private:
         {
             RpcDeferredEchoRequest __typed_req__;
             rrr::MarshalSource __req_src__(&req->m);
-            rrr::BinaryReadArchive __req_ar__(&__req_src__);
+            rrr::BinaryReadArchive __req_ar__(rrr::make_source_proxy(&__req_src__));
             __req_ar__ >> __typed_req__.val;
             auto __typed_resp__ = std::make_shared<RpcDeferredEchoResponse>();
             rrr::DeferredReply __defer__(
@@ -784,7 +784,7 @@ public:
             RpcFastPrimeResponse __typed_resp__;
             auto __reply_guard__ = __fu__->get_reply();
             rrr::MarshalSource __reply_src__(&*__reply_guard__);
-            rrr::BinaryReadArchive __reply_ar__(&__reply_src__);
+            rrr::BinaryReadArchive __reply_ar__(rrr::make_source_proxy(&__reply_src__));
             __reply_ar__ >> __typed_resp__.flag;
             return rusty::Result<RpcFastPrimeResponse, rrr::i32>::Ok(__typed_resp__);
         }
@@ -836,7 +836,7 @@ public:
             RpcFastDotProdResponse __typed_resp__;
             auto __reply_guard__ = __fu__->get_reply();
             rrr::MarshalSource __reply_src__(&*__reply_guard__);
-            rrr::BinaryReadArchive __reply_ar__(&__reply_src__);
+            rrr::BinaryReadArchive __reply_ar__(rrr::make_source_proxy(&__reply_src__));
             __reply_ar__ >> __typed_resp__.v;
             return rusty::Result<RpcFastDotProdResponse, rrr::i32>::Ok(__typed_resp__);
         }
@@ -889,7 +889,7 @@ public:
             RpcFastAddResponse __typed_resp__;
             auto __reply_guard__ = __fu__->get_reply();
             rrr::MarshalSource __reply_src__(&*__reply_guard__);
-            rrr::BinaryReadArchive __reply_ar__(&__reply_src__);
+            rrr::BinaryReadArchive __reply_ar__(rrr::make_source_proxy(&__reply_src__));
             __reply_ar__ >> __typed_resp__.a_add_b;
             return rusty::Result<RpcFastAddResponse, rrr::i32>::Ok(__typed_resp__);
         }
@@ -990,7 +990,7 @@ public:
             RpcFastVecResponse __typed_resp__;
             auto __reply_guard__ = __fu__->get_reply();
             rrr::MarshalSource __reply_src__(&*__reply_guard__);
-            rrr::BinaryReadArchive __reply_ar__(&__reply_src__);
+            rrr::BinaryReadArchive __reply_ar__(rrr::make_source_proxy(&__reply_src__));
             __reply_ar__ >> __typed_resp__.v;
             return rusty::Result<RpcFastVecResponse, rrr::i32>::Ok(__typed_resp__);
         }
@@ -1042,7 +1042,7 @@ public:
             RpcPrimeResponse __typed_resp__;
             auto __reply_guard__ = __fu__->get_reply();
             rrr::MarshalSource __reply_src__(&*__reply_guard__);
-            rrr::BinaryReadArchive __reply_ar__(&__reply_src__);
+            rrr::BinaryReadArchive __reply_ar__(rrr::make_source_proxy(&__reply_src__));
             __reply_ar__ >> __typed_resp__.flag;
             return rusty::Result<RpcPrimeResponse, rrr::i32>::Ok(__typed_resp__);
         }
@@ -1094,7 +1094,7 @@ public:
             RpcDotProdResponse __typed_resp__;
             auto __reply_guard__ = __fu__->get_reply();
             rrr::MarshalSource __reply_src__(&*__reply_guard__);
-            rrr::BinaryReadArchive __reply_ar__(&__reply_src__);
+            rrr::BinaryReadArchive __reply_ar__(rrr::make_source_proxy(&__reply_src__));
             __reply_ar__ >> __typed_resp__.v;
             return rusty::Result<RpcDotProdResponse, rrr::i32>::Ok(__typed_resp__);
         }
@@ -1147,7 +1147,7 @@ public:
             RpcAddResponse __typed_resp__;
             auto __reply_guard__ = __fu__->get_reply();
             rrr::MarshalSource __reply_src__(&*__reply_guard__);
-            rrr::BinaryReadArchive __reply_ar__(&__reply_src__);
+            rrr::BinaryReadArchive __reply_ar__(rrr::make_source_proxy(&__reply_src__));
             __reply_ar__ >> __typed_resp__.a_add_b;
             return rusty::Result<RpcAddResponse, rrr::i32>::Ok(__typed_resp__);
         }
@@ -1344,7 +1344,7 @@ public:
             RpcDeferredEchoResponse __typed_resp__;
             auto __reply_guard__ = __fu__->get_reply();
             rrr::MarshalSource __reply_src__(&*__reply_guard__);
-            rrr::BinaryReadArchive __reply_ar__(&__reply_src__);
+            rrr::BinaryReadArchive __reply_ar__(rrr::make_source_proxy(&__reply_src__));
             __reply_ar__ >> __typed_resp__.result;
             return rusty::Result<RpcDeferredEchoResponse, rrr::i32>::Ok(__typed_resp__);
         }
